Create Resources folders and move resources into them.
